backend/requirements.txt
# EPISCAN Backend Dependencies (Updated for Python 3.13)
# ======================================================

fastapi>=0.115.0
uvicorn[standard]>=0.31.0
supabase>=2.11.0
python-dotenv>=1.0.1
pydantic>=2.10.0
python-multipart>=0.0.12